fix failing test
max ID was a clever way to avoid a full scan maybe, but doesnt' work with refresh database and pgsql. For sqlite it was fine because i got a brand new in-mem database for every test. Now that I have pgsql, rolling back the transaction between tests removes data but does not reset incrementing IDs