Switch to postgres (#12)
* Update connection

* fix: foreignUlid not id

In sqlite this was okay because everything is treated as text in sqlite I believe

* example more CI friendly

* Only seed locally

* remove debug code

* use postgres aggregate

* pgsql for github ci

* update fly.toml to use pg

* Update test values

* replace mysql from connection defaults

* pint

* assertSeeInOrder preferred

I only care about the visible text content not the tags when making these assertions

* fix failing test

max ID was a clever way to avoid a full scan maybe, but doesnt' work with refresh database and pgsql. For sqlite it was fine because i got a brand new in-mem database for every test. Now that I have pgsql, rolling back the transaction between tests removes data but does not reset incrementing IDs

* lazy load potentially slow count query

* phpstan: no need for ??

expression is not nullable. will return 0 if no rows