upgrade prettier (#18)
* Fix: Remove stray header closing tag

* Move prettier config to .prettierrc

This is primarily so that PHPStorm will automatically pick it up and honor it. It is not parsing it out of package.json.

* Remove duplicated config

It's now in .prettierrc

* Upgrade prettier

- Get latest prettier
- Use prettier-plugin-blade which now supports prettier v3
- Use prettier-plugin-tailwindcss to separately sort TW classes

* Prettify

Looks like lots of changes. The switch of blade plugins appears to have created quite a few changes.

* Ensure process code 1 if prettier fails

Not sure why but I was getting 0 after a warning without this.