add healthchecks for web and cron
Cron is basically the same as the queue:work healthcheck. But i realized
that i configured it to be more like a readiness probe with the very
aggressive every 1s interval. so i am using 30s here.

However, I am leaving database at 1s because I need database to be ready
before anything else can start. (I use depends_on for this). I don't
want to sit around for 30s with a ready database that docker doesn't
know is ready before it will start up my webserver.

Maybe I should just remove the dependency actually for the webserver
though instead. that would be a different possibly better way to fix it.
like if the db is down do i actually want my webserver to be down?
probably not. better to just start?