Add stale-while-revalidate headers to home page
The goal is to serve index.html from the edge and then in the background update the cloud flare cache

The test is failing right now. This is because there is a livewire component on the home page which is overriding caching behavior. I think it has something to do with a default back-button browser history setting