Remove static asset caching
The build assets are hashed. But just blankyly caching for a year based on file ending is more dangerous. Then I have to remember to change the name of a file every time I want to update it. For instance, with things like resume.pdf.

I can always add that in later if i want to